Chore: raise PHPStan to level 7, fix union and array-offset issues
- Bump phpstan.neon level 6 → 7
- Cast strtotime() to int where date() requires int|null (9 occurrences
  across AddProjectComment, GetProject, GetUserActivity, SqlAdminRepository)
- Annotate private growth methods and getMemberGrowth() with array-shape
  @return tags so PHPStan can verify the shape contract (SqlAdminRepository)
- Narrow findRegistrationsByUserId() return via @var cast to satisfy the
  interface shape without changing runtime behaviour (SqlEventRepository)
- Fix pre-existing bug in ForumController::likePost — arguments to
  LikeForumPostInput were swapped/surplus; use-case only needs postId string

+++ src/Application/Project/AddProjectComment/AddProjectComment.php
@@ -51,7 +51,7 @@ final class AddProjectComment
             'avatar_color'    => $comment->avatarColor(),
             'content'         => htmlspecialchars($comment->content()),
             'likes'           => 0,
-            'timestamp'       => date('F j, Y, H:i', strtotime($now)),
+            'timestamp'       => date('F j, Y, H:i', (int) strtotime($now)),
         ]);
     }
